[SITE-6015] Fix Codacy: split test class, suppress unavoidable static access
